requirement.txt
asgiref==3.3.1
bce-python-sdk==0.8.59
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
django-filters==0.2.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
future==0.18.2
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
loguru==0.5.3
MarkupSafe==1.1.1
packaging==20.9
pycryptodome==3.10.1
pydantic==1.8.1
PyJWT==1.7.1
PyMySQL==1.0.2
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
uvicorn==0.11.8
uvloop==0.14.0
